Build: (d0c10fb) I've made two important changes to improve the contrast:
Removed the opacity-90 from the text-white class, which was reducing the contrast of the text. White text should be fully opaque for maximum contrast.

Changed the background from bg-blue-600 to bg-blue-800 which is a darker shade of blue. This provides better contrast with white text. The color contrast ratio will now be higher and meet WCAG accessibility guidelines.